Add user authentication

+++ routes/web.php
@@ -1,7 +1,8 @@
 <?php
 
-use App\Http\Controllers\ListingController;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\UserController;
+use App\Http\Controllers\ListingController;
 
 /*
 |--------------------------------------------------------------------------
@@ -18,22 +19,37 @@ use Illuminate\Support\Facades\Route;
 Route::get('/', [ListingController::class, 'index']);
 
 // Form to create new list
-Route::get('/listing/create', [ListingController::class, 'create']);
+Route::get('/listing/create', [ListingController::class, 'create'])->middleware('auth');
 
 // Store listing data
-Route::post('/listing', [ListingController::class, 'store']);
+Route::post('/listing', [ListingController::class, 'store'])->middleware('auth');
 
 // Show Edit Form
-Route::get('/listing/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listing/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 // Update Listing
-Route::put('/listing/{listing}', [ListingController::class, 'update']);
+Route::put('/listing/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 // Delete Listing
-Route::delete('/listing/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listing/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
 // Single listing
 Route::get('/listing/{listing}', [ListingController::class, 'show']);
+
+// Show User Register/Create Form
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
+
+// Create New User
+Route::post('/users', [UserController::class, 'store']);
+
+// Log User Out
+Route::post('/logout', [UserController::class, 'logout']);
+
+// Show Login Form
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+
+// Log In User
+Route::post('/users/authenticate', [UserController::class, 'authenticate']);
 
 // Common Resource Routes:
 // index - Show all listings
